codec aware mp4 muxer
