+++ src/main/java/org/jcodec/containers/mp4/muxer/MP4MuxerTrack.java
@@ -1,36 +1,22 @@
 package org.jcodec.containers.mp4.muxer;
 
+import static org.jcodec.common.Ints.checkedCast;
+import static org.jcodec.common.Preconditions.checkState;
+
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Date;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 
-import org.jcodec.codecs.aac.ADTSParser;
-import org.jcodec.codecs.h264.H264Utils;
-import org.jcodec.codecs.h264.io.model.SeqParameterSet;
-import org.jcodec.codecs.mpeg4.mp4.EsdsBox;
-import org.jcodec.common.Assert;
-import org.jcodec.common.AudioFormat;
-import org.jcodec.common.Codec;
 import org.jcodec.common.IntArrayList;
 import org.jcodec.common.LongArrayList;
-import org.jcodec.common.VideoCodecMeta;
-import org.jcodec.common.io.NIOUtils;
 import org.jcodec.common.io.SeekableByteChannel;
-import org.jcodec.common.model.ColorSpace;
 import org.jcodec.common.model.Packet;
-import org.jcodec.common.model.Packet.FrameType;
 import org.jcodec.common.model.Rational;
 import org.jcodec.common.model.Size;
 import org.jcodec.common.model.Unit;
 import org.jcodec.containers.mp4.MP4TrackType;
-import org.jcodec.containers.mp4.boxes.AudioSampleEntry;
 import org.jcodec.containers.mp4.boxes.Box;
 import org.jcodec.containers.mp4.boxes.ChunkOffsets64Box;
 import org.jcodec.containers.mp4.boxes.CompositionOffsetsBox;
@@ -44,7 +30,6 @@ import org.jcodec.containers.mp4.boxes.MediaHeaderBox;
 import org.jcodec.containers.mp4.boxes.MediaInfoBox;
 import org.jcodec.containers.mp4.boxes.MovieHeaderBox;
 import org.jcodec.containers.mp4.boxes.NodeBox;
-import org.jcodec.containers.mp4.boxes.PixelAspectExt;
 import org.jcodec.containers.mp4.boxes.SampleDescriptionBox;
 import org.jcodec.containers.mp4.boxes.SampleEntry;
 import org.jcodec.containers.mp4.boxes.SampleSizesBox;
@@ -55,7 +40,6 @@ import org.jcodec.containers.mp4.boxes.TimeToSampleBox;
 import org.jcodec.containers.mp4.boxes.TimeToSampleBox.TimeToSampleEntry;
 import org.jcodec.containers.mp4.boxes.TrackHeaderBox;
 import org.jcodec.containers.mp4.boxes.TrakBox;
-import org.jcodec.containers.mp4.boxes.VideoSampleEntry;
 
 /**
  * This class is part of JCodec ( www.jcodec.org ) This software is distributed
@@ -64,18 +48,7 @@ import org.jcodec.containers.mp4.boxes.VideoSampleEntry;
  * @author The JCodec project
  * 
  */
-public class FramesMP4MuxerTrack extends AbstractMP4MuxerTrack {
-
-    private static Map<Codec, String> codec2fourcc = new HashMap<Codec, String>();
-
-    static {
-        codec2fourcc.put(Codec.H264, "avc1");
-        codec2fourcc.put(Codec.AAC, "mp4a");
-        codec2fourcc.put(Codec.PRORES, "apch");
-        codec2fourcc.put(Codec.JPEG, "mjpg");
-        codec2fourcc.put(Codec.PNG, "png ");
-        codec2fourcc.put(Codec.V210, "v210");
-    }
+public class MP4MuxerTrack extends AbstractMP4MuxerTrack {
 
     private List<TimeToSampleEntry> sampleDurations;
     private long sameDurCount = 0;
@@ -97,16 +70,8 @@ public class FramesMP4MuxerTrack extends AbstractMP4MuxerTrack {
     private boolean allIframes = true;
     private TimecodeMP4MuxerTrack timecodeTrack;
     private SeekableByteChannel out;
-    private Codec codec;
 
-    // SPS/PPS lists when h.264 is stored, otherwise these lists are not used.
-    private List<ByteBuffer> spsList = new ArrayList<ByteBuffer>();
-    private List<ByteBuffer> ppsList = new ArrayList<ByteBuffer>();
-
-    // ADTS header used to construct audio sample entry for AAC
-    private ADTSParser.Header adtsHeader;
-
-    public FramesMP4MuxerTrack(SeekableByteChannel out, int trackId, MP4TrackType type, Codec codec) {
+    public MP4MuxerTrack(SeekableByteChannel out, int trackId, MP4TrackType type) {
         super(trackId, type);
         this.sampleDurations = new ArrayList<TimeToSampleEntry>();
         this.chunkOffsets = LongArrayList.createLongArrayList();
@@ -116,30 +81,11 @@ public class FramesMP4MuxerTrack extends AbstractMP4MuxerTrack {
         
         this.out = out;
 
-        this.codec = codec;
-
         setTgtChunkDuration(new Rational(1, 1), Unit.FRAME);
     }
 
+    @Override
     public void addFrame(Packet pkt) throws IOException {
-        if (codec == Codec.H264) {
-            ByteBuffer result = pkt.getData();
-            
-            if (pkt.frameType == FrameType.UNKOWN) {
-                pkt.setFrameType(H264Utils.isByteBufferIDRSlice(result) ? FrameType.KEY : FrameType.INTER);
-            }
-            
-            H264Utils.wipePSinplace(result, spsList, ppsList);
-            result = H264Utils.encodeMOVPacket(result);
-            pkt = Packet.createPacketWithData(pkt, result);
-        } else if (codec == Codec.AAC) {
-            ByteBuffer result = pkt.getData();
-            adtsHeader = ADTSParser.read(result);
-//            System.out.println(String.format("crc_absent: %d, num_aac_frames: %d, size: %d, remaining: %d, %d, %d, %d",
-//                    adtsHeader.getCrcAbsent(), adtsHeader.getNumAACFrames(), adtsHeader.getSize(), result.remaining(),
-//                    adtsHeader.getObjectType(), adtsHeader.getSamplingIndex(), adtsHeader.getChanConfig()));
-            pkt = Packet.createPacketWithData(pkt, result);
-        }
         addFrameInternal(pkt, 1);
         processTimecode(pkt);
     }
@@ -149,11 +95,7 @@ public class FramesMP4MuxerTrack extends AbstractMP4MuxerTrack {
             throw new IllegalStateException("The muxer track has finished muxing");
 
         if (_timescale == NO_TIMESCALE_SET) {
-            if (adtsHeader != null) {
-                _timescale = adtsHeader.getSampleRate();
-            } else {
-                _timescale = pkt.getTimescale();
-            }
+            _timescale = pkt.getTimescale();
         }
         
         if (_timescale != pkt.getTimescale()) {
@@ -161,10 +103,6 @@ public class FramesMP4MuxerTrack extends AbstractMP4MuxerTrack {
             pkt.setDuration((pkt.getPts() * _timescale) / pkt.getDuration());
         }
         
-        if (adtsHeader != null) {
-            pkt.setDuration(1024);
-        }
-
         if(type == MP4TrackType.VIDEO) {
             long compositionOffset = pkt.getPts() - ptsEstimate;
             if (compositionOffset != lastCompositionOffset) {
@@ -211,7 +149,7 @@ public class FramesMP4MuxerTrack extends AbstractMP4MuxerTrack {
     }
 
     private void outChunkIfNeeded(int entryNo) throws IOException {
-        Assert.assertTrue(tgtChunkDurationUnit == Unit.FRAME || tgtChunkDurationUnit == Unit.SEC);
+        checkState(tgtChunkDurationUnit == Unit.FRAME || tgtChunkDurationUnit == Unit.SEC);
 
         if (tgtChunkDurationUnit == Unit.FRAME
                 && curChunk.size() * tgtChunkDuration.getDen() == tgtChunkDuration.getNum()) {
@@ -243,20 +181,9 @@ public class FramesMP4MuxerTrack extends AbstractMP4MuxerTrack {
         curChunk.clear();
     }
 
+    @Override
     protected Box finish(MovieHeaderBox mvhd) throws IOException {
-        if (finished)
-            throw new IllegalStateException("The muxer track has finished muxing");
-        if (getEntries().isEmpty()) {
-            if (codec == Codec.H264) {
-                SeqParameterSet sps = SeqParameterSet.read(spsList.get(0).duplicate());
-                Size size = H264Utils.getPicSize(sps);
-                VideoCodecMeta meta = org.jcodec.common.VideoCodecMeta.createSimpleVideoCodecMeta(size, ColorSpace.YUV420);
-                addVideoSampleEntry(meta);
-            } else {
-                throw new RuntimeException("Sample entry missing not supported for anything other then H.264");
-            }
-        }
-        setCodecPrivateIfNeeded();
+        checkState(finished, "The muxer track has finished muxing");
 
         outChunk(lastEntry);
 
@@ -268,7 +195,7 @@ public class FramesMP4MuxerTrack extends AbstractMP4MuxerTrack {
         TrakBox trak = TrakBox.createTrakBox();
         Size dd = getDisplayDimensions();
         TrackHeaderBox tkhd = TrackHeaderBox.createTrackHeaderBox(trackId,
-                ((long) mvhd.getTimescale() * trackTotalDuration) / _timescale, dd.getWidth(), dd.getHeight(),
+                (mvhd.getTimescale() * trackTotalDuration) / _timescale, dd.getWidth(), dd.getHeight(),
                 new Date().getTime(), new Date().getTime(), 1.0f, (short) 0, 0,
                 new int[] { 0x10000, 0, 0, 0, 0x10000, 0, 0, 0, 0x40000000 });
         tkhd.setFlags(0xf);
@@ -305,13 +232,6 @@ public class FramesMP4MuxerTrack extends AbstractMP4MuxerTrack {
             stbl.add(SyncSamplesBox.createSyncSamplesBox(iframes.toArray()));
 
         return trak;
-    }
-
-    void addVideoSampleEntry(VideoCodecMeta meta) {
-        SampleEntry se = VideoSampleEntry.videoSampleEntry(codec2fourcc.get(codec), meta.getSize(), "JCodec");
-        if (meta.getPixelAspectRatio() != null)
-            se.add(PixelAspectExt.createPixelAspectExt(meta.getPixelAspectRatio()));
-        addSampleEntry(se);
     }
 
     private void putCompositionOffsets(NodeBox stbl) {
@@ -363,6 +283,7 @@ public class FramesMP4MuxerTrack extends AbstractMP4MuxerTrack {
         return min;
     }
 
+    @Override
     public long getTrackTotalDuration() {
         return trackTotalDuration;
     }
@@ -380,73 +301,5 @@ public class FramesMP4MuxerTrack extends AbstractMP4MuxerTrack {
 
     public void setTimecode(TimecodeMP4MuxerTrack timecodeTrack) {
         this.timecodeTrack = timecodeTrack;
-    }
-
-    public void setCodecPrivateIfNeeded() {
-        if (codec == Codec.H264) {
-            getEntries().get(0).add(H264Utils.createAvcCFromPS(selectUnique(spsList), selectUnique(ppsList), 4));
-        } else if (codec == Codec.AAC) {
-            getEntries().get(0).add(EsdsBox.fromADTS(adtsHeader));
-        }
-    }
-
-    private static class ByteArrayWrapper {
-        private byte[] bytes;
-
-        public ByteArrayWrapper(ByteBuffer bytes) {
-            this.bytes = NIOUtils.toArray(bytes);
-        }
-
-        public ByteBuffer get() {
-            return ByteBuffer.wrap(bytes);
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (!(obj instanceof ByteArrayWrapper))
-                return false;
-            return Arrays.equals(bytes, ((ByteArrayWrapper) obj).bytes);
-        }
-
-        @Override
-        public int hashCode() {
-            return Arrays.hashCode(bytes);
-        }
-    }
-
-    private List<ByteBuffer> selectUnique(List<ByteBuffer> bblist) {
-        Set<ByteArrayWrapper> all = new HashSet<ByteArrayWrapper>();
-        for (ByteBuffer byteBuffer : bblist) {
-            all.add(new ByteArrayWrapper(byteBuffer));
-        }
-        List<ByteBuffer> result = new ArrayList<ByteBuffer>();
-        for (ByteArrayWrapper bs : all) {
-            result.add(bs.get());
-        }
-        return result;
-    }
-
-    void addAudioSampleEntry(AudioFormat format) {
-        AudioSampleEntry ase = AudioSampleEntry.compressedAudioSampleEntry(codec2fourcc.get(codec), (short) 1, (short) 16,
-                format.getChannels(), format.getSampleRate(), 0, 0, 0);
-
-        addSampleEntry(ase);
-    }
-    
-    /**
-     * Returns the {@code int} value that is equal to {@code value}, if possible.
-     *
-     * @param value any value in the range of the {@code int} type
-     * @return the {@code int} value that equals {@code value}
-     * @throws IllegalArgumentException if {@code value} is greater than {@link
-     *     Integer#MAX_VALUE} or less than {@link Integer#MIN_VALUE}
-     */
-    public static int checkedCast(long value) {
-      int result = (int) value;
-      if (result != value) {
-        // don't use checkArgument here, to avoid boxing
-        throw new IllegalArgumentException("Out of range: " + value);
-      }
-      return result;
     }
 }
